Refactor code to implement mean reversion strategy
Modified the backtesting code to use the new MeanReversionBacktester and MeanReversionStrategy classes. In addition, the data provider file was deleted and replaced with updated data providers that are now incorporated directly into the backtesting code. To accommodate these changes, references to the old data provider and backtesting classes were updated across multiple files.